+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.7.0
 ### Features
 * add option caseSensitiveAutocomplete default to true [#332](https://github.com/jcubic/jquery.terminal/issues/332)
+* expose Stack/Cycle/History in $.terminal so they can be tested
 ### Bugs
 * fix jumping of terminal when created one after another and changing the one that have focus in Edge
 * fix issue that all terminals was enabled not the last one created